requirements.txt
streamlit==1.32.2
groq==0.4.3
langchain-groq==0.0.5
langchain==0.1.12
langchain-huggingface==0.0.9
langchain-community==0.0.28
pinecone-client==3.1.0
sentence-transformers==2.5.1